Projects with same prefix fix
If a project named "aaa" is imported, then you import
another named "aaa-bbb" and then try to git-tf clone "aaa-bbb"
it failed saying that you must be in a local path where "aaa"
was.

Thats because:

/path/to/workspace/aaa
/path/to/workspace/aaa-bbb

have a common os.path.commonprefix. And every run of clone in
/path/to/workspace/aaa-bbb keeps complaining that you need to be
in /path/to/workspace/aaa.