Bad substitution error.
When pushing or fetching commits with "${ }" in message the linux/unix
shell fails with the message "Bad substitution error". This is because
"${ }" has a special meaning in shell strings. For example:

$ echo "${ }"
bash: ${ }: bad substitution

Note: this fix works for linux. Somebody on Windows should try that this
don't cause any problems. If yes -- it should be included in the changes
Nodirt is making in the win branch.